feat(collections): stabilize Iterable input for zip()
